+++ packages/guide/src/app/core/guards/only-in-beta.guard.ts
@@ -1,14 +1,15 @@
-import { ActivatedRouteSnapshot, CanActivateFn, createUrlTreeFromSnapshot } from "@angular/router";
+import { ActivatedRouteSnapshot, CanActivateFn, Router } from "@angular/router";
 import { inject } from "@angular/core";
 import { SettingsService } from "../../shared/services/settings.service";
 
 export const onlyInBetaGuard: CanActivateFn = (route: ActivatedRouteSnapshot,) => {
     const settings = inject(SettingsService);
+    const router = inject(Router);
 
     if (settings.getSettings().useBeta) {
         return true
     } else {
-        return createUrlTreeFromSnapshot(route, ['/only-in-beta'])
+        return router.createUrlTree(['/only-in-beta'])
     }
 
 };