+++ packages/pak-files-parser/src/interfaces/raw-data-interfaces/raw-treasure-hunt.type.ts
@@ -1,0 +1,17 @@
+export type RawTreasureHunt = {
+    "treasureId": number,
+    "mapItemData": {
+        "data": {
+            "DataTable": {
+                "ObjectName": string
+                "ObjectPath": string
+            },
+            "RowName": string
+        },
+        "itemID":string
+    },
+    "treasureLocationImage": {
+        "AssetPathName": string
+        "SubPathString": string
+    }
+}