+++ packages/guide/src/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Changelog
 
+### TBD
+
+#### Added
+
+- \[Beta] Added ocean produce to journal
+- \[Beta] Added treasure maps to journal
+
+#### Changed
+
+- \[Beta] Fixed missing beta notification for orchestra zones
+- Fixed link to settings in "beta only"-landing page
+
 ### 0.19.0 2024-05-22
 
 #### Added